+++ resources/views/components/streamgraph.blade.php
@@ -1,0 +1,9 @@
+@props([
+    'series' => [],
+    'categories' => [],
+    'legend' => true,
+])
+
+<x-chart::base component="streamgraph" :option="\WireCharts\Support\Option::streamgraph([
+    'series' => $series, 'categories' => $categories, 'legend' => $legend,
+])" {{ $attributes }} />